requirements-streamlit.txt
streamlit==1.31.0
plotly==5.18.0
pandas==2.2.2
psycopg2-binary==2.9.9
numpy==1.26.3
python-dateutil==2.8.2
